requirements.txt
tensorflow-cpu
numpy
fastapi==0.70.0
Pillow==8.4.0
uvicorn==0.15.0
python-multipart
